+++ CHANGELOG.md
@@ -6,7 +6,11 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
-## [408.2] - beta
+## [409.0] - beta
+
+* No significant changes
+
+## [408.2] - 2026.04.13
 
 ### Added